requirements.txt
requests==2.18.4
pycryptodome==3.9.7
PyQt5==5.14.2
aiohttp==3.6.2
quamash==0.6.1
urllib3==1.22
m3u8==0.9.0